+++ ament_clang_format/package.xml
@@ -14,7 +14,7 @@
   <buildtool_export_depend>ament_cmake_test</buildtool_export_depend>
   <buildtool_export_depend>clang-format</buildtool_export_depend>
 
-  <test_depend>ament_lint_cmake</test_depend>
+  <test_depend>ament_cmake_lint_cmake</test_depend>
   <test_depend>ament_pep8</test_depend>
   <test_depend>ament_pyflakes</test_depend>
 </package>
