+++ ament_cmake_cppcheck/package.xml
@@ -17,5 +17,5 @@
   <buildtool_export_depend>ament_cppcheck</buildtool_export_depend>
 
   <test_depend>ament_cmake_copyright</test_depend>
-  <test_depend>ament_lint_cmake</test_depend>
+  <test_depend>ament_cmake_lint_cmake</test_depend>
 </package>
