+++ ament_cmake_lint_cmake/CMakeLists.txt
@@ -1,16 +1,16 @@
 cmake_minimum_required(VERSION 2.8.3)
 
-project(ament_cmake_copyright NONE)
+project(ament_cmake_lint_cmake NONE)
 
 find_package(ament_cmake_core REQUIRED)
 find_package(ament_cmake_test REQUIRED)
 
 # include CMake functions of this package
-set(ament_cmake_copyright_DIR "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
-include("ament_cmake_copyright-extras.cmake" NO_POLICY_SCOPE)
+set(ament_cmake_lint_cmake_DIR "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
+include("ament_cmake_lint_cmake-extras.cmake" NO_POLICY_SCOPE)
 
 ament_package(
-  CONFIG_EXTRAS "ament_cmake_copyright-extras.cmake"
+  CONFIG_EXTRAS "ament_cmake_lint_cmake-extras.cmake"
 )
 
 install(
@@ -19,8 +19,5 @@ install(
 )
 
 if(AMENT_ENABLE_TESTING)
-  find_package(ament_cmake_lint_cmake REQUIRED)
   ament_lint_cmake()
-
-  ament_copyright("cmake")
 endif()
