+++ ament_cmake_lint_cmake/package.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>ament_cmake_copyright</name>
+  <name>ament_cmake_lint_cmake</name>
   <version>0.0.0</version>
-  <description>The CMake API for ament_copyright to check every source file contains copyright reference.</description>
+  <description>
+    The CMake API for ament_lint_cmake to lint CMake code using cmakelint.
+  </description>
   <maintainer email="dthomas@osrfoundation.org">Dirk Thomas</maintainer>
   <license>Apache License 2.0</license>
 
   <buildtool_depend>ament_cmake_core</buildtool_depend>
   <buildtool_depend>ament_cmake_test</buildtool_depend>
-  <buildtool_depend>ament_copyright</buildtool_depend>
+  <buildtool_depend>ament_lint_cmake</buildtool_depend>
 
   <buildtool_export_depend>ament_cmake_test</buildtool_export_depend>
-  <buildtool_export_depend>ament_copyright</buildtool_export_depend>
-
-  <test_depend>ament_cmake_lint_cmake</test_depend>
+  <buildtool_export_depend>ament_lint_cmake</buildtool_export_depend>
 </package>
