+++ ament_uncrustify/CMakeLists.txt
@@ -26,6 +26,9 @@ install(
 )
 
 if(AMENT_ENABLE_TESTING)
+  find_package(ament_cmake_lint_cmake REQUIRED)
+  ament_lint_cmake()
+
   set(python_files "ament_uncrustify" "bin/ament_uncrustify")
 
   find_package(ament_pep8 REQUIRED)
@@ -33,7 +36,4 @@ if(AMENT_ENABLE_TESTING)
 
   find_package(ament_pyflakes REQUIRED)
   ament_pyflakes(${python_files})
-
-  find_package(ament_lint_cmake REQUIRED)
-  ament_lint_cmake()
 endif()
